+++ test/rule_test.dart
@@ -4,9 +4,18 @@ import 'package:test/test.dart';
 void main() {
   group('Rule', () {
     test('Compare equal', () {
-      expect(Deny(source: 'foo', target: 'bar'), equals(Deny(source: 'foo', target: 'bar')));
-      expect(Allow(source: 'foo', target: 'bar'), equals(Allow(source: 'foo', target: 'bar')));
-      expect(Allow(source: 'foo', target: 'bar'), isNot(equals(Deny(source: 'foo', target: 'bar'))));
+      expect(
+        Deny(sources: ['foo'], targets: ['bar']),
+        equals(Deny(sources: ['foo'], targets: ['bar'])),
+      );
+      expect(
+        Allow(sources: ['foo'], targets: ['bar']),
+        equals(Allow(sources: ['foo'], targets: ['bar'])),
+      );
+      expect(
+        Allow(sources: ['foo'], targets: ['bar']),
+        isNot(equals(Deny(sources: ['foo'], targets: ['bar']))),
+      );
     });
   });
 }